django_workshop/requirements.txt
Django>=5.0,<6.0
djangorestframework
whitenoise
gunicorn
